lab_4/file_storage_service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
PyJWT==2.10.1
passlib==1.7.4
pip==25.0.1
pycparser==2.22
pydantic==2.11.0
pydantic_core==2.33.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.13.0
typing-inspection==0.4.0
uvicorn==0.34.0
pymongo==4.6.3
